requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
mediapipe==0.10.7
tensorflow==2.15.0
opencv-python==4.10.0.84
numpy==1.24.4
python-multipart
mysql-connector-python
bcrypt
pydantic[email]